web/requirements.txt
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
python-scrapyd-api==2.1.2
pydantic~=1.10.4
fastapi~=0.89.1
uvicorn==0.18.3
celery~=5.2.7
redis==4.4.2
python-dotenv~=0.21.1
pytz~=2022.7.1
peewee==3.15.4
psycopg2-binary==2.9.5
pandas==1.5.3
cdifflib==1.2.6
